Damage generator addon
Added damage generator which works on tensors (gpu). It can add damage to the whole batch of images very fast.

It has been integrated into the DataLoader but can be called itself as well, then it will generate mask for damage.

When called standalone you can check params of the damage.

Integrated version is currently hardcoded for ranges of damage (it is chosen randomly but ranges of ratios, widths etc. are hardcoded). I tested it to be quite well suited for generating damages in our task.