InPainter prototype_V1
Added prototype architecture using atrous multiheads with u net autoencoder.

Fixed bug with last batch not being the batch_size (most of the time it would be true)

Prepared trainnig loop as well as added file with inpainter functions usefull while working on the model (to not declare them in main).

Model is ready for training and has basic callbacks.