Baseline encoder commit
Few improvements in the functions and config changes (so split for val and test is smaller - working with unsupervised anyway so less important if even important at all)

Splitted encoder and decoder and adapted training loop for the full training of autoencoder.

Added new display (pain in the ass) of splitted autoencoder.